Cleanup of methods dealing with the SciObj byte store

+++ gmn/src/d1_gmn/app/sciobj_store.py
@@ -26,20 +26,23 @@
 - Folders are created as required in the hierarchy.
 
 """
+import d1_common.iter.stream
 
+import d1_common.const
 import hashlib
 import os
 import re
 
 import contextlib2
+import d1_common.iter
 
 import d1_gmn
 
-import d1_common.iter.file
 import d1_common.types
 import d1_common.types.exceptions
 import d1_common.util
 import d1_common.utils.filesystem
+import d1_gmn.app
 
 import django.conf
 
@@ -56,30 +59,20 @@ RELATIVE_PATH_MAGIC_HOST_STR = 'gmn-object-store'
 # Default location
 
 
-def save_in_object_store_by_file(pid, sciobj_file):
-    """Save the Science Object bytes in the ``sciobj_file`` file-like object to the
-    default location within the tree of the local SciObj store and return file_url with
-    the file location in a suitable form for storing in the DB."""
-    return save_in_object_store_by_iter(
-        pid,
-        d1_common.iter.file.FileLikeObjectIterator(
-            sciobj_file, chunk_size=django.conf.settings.NUM_CHUNK_BYTES
-        ),
-    )
-
-
-def save_in_object_store_by_iter(pid, sciobj_iter):
-    """Save the Science Object bytes in the ``sciobj_iter`` iterator object to the
-    default location within the tree of the local SciObj store and return file_url with
-    the file location in a suitable form for storing in the DB."""
-    with open_sciobj_file_by_pid(pid, write=True) as (sciobj_file, file_url):
-        for chunk_str in sciobj_iter:
-            sciobj_file.write(chunk_str)
-        return file_url
+# def save_in_object_store_by_file(pid, sciobj_file):
+#     """Save the Science Object bytes in the ``sciobj_file`` file-like object to the
+#     default location within the tree of the local SciObj store and return file_url with
+#     the file location in a suitable form for storing in the DB."""
+#     return save_in_object_store_by_iter(
+#         pid,
+#         d1_common.iter.stream.StreamIterator(
+#             sciobj_file, chunk_size=django.conf.settings.NUM_CHUNK_BYTES
+#         ),
+#     )
 
 
 @contextlib2.contextmanager
-def open_sciobj_file_by_pid(pid, write=False):
+def open_sciobj_file_by_pid_CTX(pid, write=False):
     """Open the file containing the Science Object bytes of ``pid`` in the default
     location within the tree of the local SciObj store.
 
@@ -87,41 +80,16 @@ def open_sciobj_file_by_pid(pid, write=False):
     created. Return the file handle and file_url with the file location in a suitable
     form for storing in the DB.
 
-    If nothing was written to the file, delete it.
+    If nothing was written to the file, it is deleted.
 
     """
     abs_path = get_abs_sciobj_file_path_by_pid(pid)
-    with open_sciobj_file_by_path(abs_path, write) as sciobj_file:
-        yield sciobj_file, get_rel_sciobj_file_url_by_pid(pid)
-
-
-# Custom location
-
-
-def save_in_custom_location_by_file(abs_path, sciobj_file):
-    """Save the Science Object bytes in the ``sciobj_file`` file-like object to the
-    custom location ``abs_path`` in the local filesystem and return file_url with the
-    file location in a suitable form for storing in the DB."""
-    return save_in_custom_location_by_iter(
-        abs_path,
-        d1_common.iter.file.FileLikeObjectIterator(
-            sciobj_file, chunk_size=django.conf.settings.NUM_CHUNK_BYTES
-        ),
-    )
-
-
-def save_in_custom_location_by_iter(abs_path, sciobj_iter):
-    """Save the Science Object bytes in the ``sciobj_file`` iterator object to the
-    custom location ``abs_path`` in the local filesystem and return file_url with the
-    file location in a suitable form for storing in the DB."""
-    with open_sciobj_file_by_path(abs_path, True) as sciobj_file:
-        for chunk_str in sciobj_iter:
-            sciobj_file.write(chunk_str)
-    return get_abs_sciobj_file_url(abs_path)
+    with open_sciobj_file_by_path_CTX(abs_path, write) as sciobj_file:
+        yield sciobj_file
 
 
 @contextlib2.contextmanager
-def open_sciobj_file_by_path(abs_path, write=False):
+def open_sciobj_file_by_path_CTX(abs_path, write=False):
     """Open the file containing the Science Object bytes at the custom location
     ``abs_path`` in the local filesystem.
 
@@ -138,8 +106,108 @@ def open_sciobj_file_by_path(abs_path, write=False):
         with open(abs_path, 'wb' if write else 'rb') as sciobj_file:
             yield sciobj_file
     finally:
-        if not os.path.getsize(abs_path):
+        if os.path.exists(abs_path) and not os.path.getsize(abs_path):
             os.unlink(abs_path)
+
+def get_sciobj_iter_by_url(sciobj_url):
+    abs_path = get_abs_sciobj_file_path_by_url(sciobj_url)
+    return d1_common.iter.stream.StreamIterator(
+        open_sciobj_file_by_path_PLAIN(abs_path)
+    )
+
+# @contextlib2.contextmanager
+def get_sciobj_byte_iterator_by_url(sciobj_url):
+    with open_sciobj_file_by_pid_PLAIN(get_abs_sciobj_file_path_by_url(sciobj_url)) as sciobj_file:
+        with d1_common.iter.stream.StreamIterator(sciobj_file) as sciobj_iter:
+            yield sciobj_iter
+
+def get_sciobj_iter_by_pid(pid):
+    abs_path = get_abs_sciobj_file_path_by_pid(pid)
+    return d1_common.iter.stream.StreamIterator(
+        open_sciobj_file_by_path_PLAIN(abs_path)
+    )
+
+
+# @contextlib2.contextmanager
+# def get_sciobj_byte_iterator_by_pid(pid):
+#     with open_sciobj_file_by_pid_PLAIN(pid) as sciobj_file:
+#         with d1_common.iter.stream.StreamIterator(sciobj_file) as sciobj_iter:
+#             yield sciobj_iter
+
+
+# def get_sciobj_byte_iterator_by_url_not_context(sciobj_url):
+#     abs_path = get_abs_sciobj_file_path_by_url(sciobj_url)
+#     sciobj_file = open_sciobj_file_by_path_PLAIN(abs_path)
+#     return d1_common.iter.stream.StreamIterator(sciobj_file)
+
+
+
+# def save_in_object_store_by_iter(pid, sciobj_iter):
+#     """Save the Science Object bytes in the ``sciobj_iter`` iterator object to the
+#     default location within the tree of the local SciObj store and return file_url with
+#     the file location in a suitable form for storing in the DB."""
+#     with open_sciobj_file_by_pid(pid, write=True) as (sciobj_file, file_url):
+#         for chunk_str in sciobj_iter:
+#             sciobj_file.write(chunk_str)
+#         return file_url
+
+
+# Custom location
+
+
+# def save_in_custom_location_by_file(abs_path, sciobj_file):
+#     """Save the Science Object bytes in the ``sciobj_file`` file-like object to the
+#     custom location ``abs_path`` in the local filesystem and return file_url with the
+#     file location in a suitable form for storing in the DB."""
+#     return save_in_custom_location_by_iter(
+#         abs_path,
+#         d1_common.iter.stream.StreamIterator(
+#             sciobj_file, chunk_size=django.conf.settings.NUM_CHUNK_BYTES
+#         ),
+#     )
+
+
+# def save_in_custom_location_by_iter(abs_path, sciobj_iter):
+#     """Save the Science Object bytes in the ``sciobj_file`` iterator object to the
+#     custom location ``abs_path`` in the local filesystem and return file_url with the
+#     file location in a suitable form for storing in the DB."""
+#     with open_sciobj_file_by_path_CTX(abs_path, True) as sciobj_file:
+#         for chunk_str in sciobj_iter:
+#             sciobj_file.write(chunk_str)
+#     return get_abs_sciobj_file_url(abs_path)
+
+
+# def save_sciobj_bytes_from_str(map_xml, sciobj_path):
+#     d1_common.utils.filesystem.create_missing_directories_for_file(sciobj_path)
+#     with open(sciobj_path, 'wb') as f:
+#         f.write(map_xml)
+
+
+def open_sciobj_file_by_pid_PLAIN(pid, write=False):
+    """Open the file containing the Science Object bytes at the custom location
+    ``abs_path`` in the local filesystem for read.
+    """
+    abs_path = get_abs_sciobj_file_path_by_pid(pid)
+    if write:
+        d1_common.utils.filesystem.create_missing_directories_for_file(abs_path)
+    return open_sciobj_file_by_path_PLAIN(abs_path, write)
+
+
+def open_sciobj_file_by_path_PLAIN(abs_path, write=False):
+    """Open a SciObj file for read or write. If opened for write, create any missing
+    directories. For a SciObj stored in the default SciObj store, the path includes
+    the PID hash based directory levels.
+
+    This is the only method in GMN that opens SciObj files, so can be modified to
+    customize the SciObj storage locations and can be mocked for testing.
+
+    Note that when a SciObj is created by a client via MNStorage.create(), Django
+    streams the SciObj bytes to a temporary file or memory location as set by
+    ``FILE_UPLOAD_TEMP_DIR`` and related settings.
+    """
+    if write:
+        d1_common.utils.filesystem.create_missing_directories_for_file(abs_path)
+    return open(abs_path, 'wb' if write else 'rb')
 
 
 def get_rel_sciobj_file_path(pid):
